remove warning …

+++ saytheword/Classes/Model/WordInfo/WordInfo.m
@@ -24,7 +24,7 @@
         dummyString = __initString;
         if (dummyString.length<11)
         {
-            int k = 11-dummyString.length;
+            int k = 11-(int)dummyString.length;
             for (int i=0;i<k;i++)
             {
                 dummyString = [dummyString stringByAppendingString:@"A"];
